changed id type from integer to long

+++ src/main/java/uit/java/mysql/repositories/AnnotationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import uit.java.mysql.databases.Annotation;
 
 @Repository
-public interface AnnotationRepository extends JpaRepository<Annotation, Integer> {
+public interface AnnotationRepository extends JpaRepository<Annotation, Long> {
 }
